Revert "prettier applied to all backend files"
This reverts commit e88b533de89b1576ece17a99573493b368be5532.

+++ backend/src/controller/studio.js
@@ -1,12 +1,4 @@
-const {
-    save,
-    saveImage,
-    saveImages,
-    getAll,
-    getOneByStudioId,
-    getOwnByUserId,
-    deleteById,
-} = require('../service/studio');
+const { save, saveImage, saveImages, getAll, getOneByStudioId, getOwnByUserId, deleteById } = require('../service/studio');
 
 exports.save = async (req, res, next) => {
     try {
@@ -16,7 +8,7 @@ exports.save = async (req, res, next) => {
     } catch (error) {
         next(error);
     }
-};
+}
 
 exports.saveImage = async (req, res, next) => {
     try {
@@ -27,7 +19,7 @@ exports.saveImage = async (req, res, next) => {
     } catch (error) {
         next(error);
     }
-};
+}
 
 exports.saveImages = async (req, res, next) => {
     try {
@@ -37,7 +29,7 @@ exports.saveImages = async (req, res, next) => {
     } catch (error) {
         next(error);
     }
-};
+}
 
 exports.getAll = async (req, res, next) => {
     try {
@@ -47,17 +39,17 @@ exports.getAll = async (req, res, next) => {
     } catch (error) {
         next(error);
     }
-};
+}
 
 exports.getOne = async (req, res, next) => {
-    try {
-        const { status, payload } = await getOneByStudioId(req.params.id);
+	try {
+		const { status, payload } = await getOneByStudioId(req.params.id);
 
-        res.status(status).json(payload);
-    } catch (error) {
-        next(error);
-    }
-};
+		res.status(status).json(payload);
+	} catch (error) {
+		next(error);
+	}
+}
 
 exports.getOwn = async (req, res, next) => {
     try {
@@ -67,14 +59,14 @@ exports.getOwn = async (req, res, next) => {
     } catch (error) {
         next(error);
     }
-};
+}
 
 exports.deleteOne = async (req, res, next) => {
-    try {
-        const { status, payload } = await deleteById(req.user.id);
+	try {
+		const { status, payload } = await deleteById(req.user.id);
 
-        res.status(status).json(payload);
-    } catch (error) {
-        next(error);
-    }
-};
+		res.status(status).json(payload);
+	} catch (error) {
+		next(error);
+	}
+}
